[SHANIA][ADD/UPDATE] Added new module for new navigation and edit nav
